fix some more signed shifts

+++ stb_vorbis.c
@@ -1293,7 +1293,7 @@ static uint32 get32(vorb *f)
    x = get8(f);
    x += get8(f) << 8;
    x += get8(f) << 16;
-   x += get8(f) << 24;
+   x += (uint32) get8(f) << 24;
    return x;
 }
